+++ src/main/java/com/quaxt/mcc/tacky/CopyToOffset.java
@@ -1,4 +1,4 @@
 package com.quaxt.mcc.tacky;
 
-public record CopyToOffset(ValIr src, VarIr dst, int offset) implements InstructionIr {
+public record CopyToOffset(ValIr src, VarIr dst, long offset) implements InstructionIr {
 }
